v3_core: Eliminate double processing of patterns
Internally in the v3_core pass, an #imatch{} record represents
a match expression:

  Pattern = Expression

If Pattern is a single, unbound variable, #imatch{} will be
rewritten to #iset{}; otherwise it will be rewritten to #icase{}.

To determine how #imatch{} should be translated, the pattern is
processed using upattern/3. The return value from upattern/3 is thrown
away (after having been used for determing how the #imatch{} record
should be translated).

That means that every pattern in an #imatch{} is processed twice,
which is wasteful.

We can easily avoid the double processing of patterns by
introducing a new helper function that determines whether the
pattern is a new variable.