[ct_netconfc] Fix XML parsing when multiple messages in package
If a ssh package contained more than one netconf end tag, then the
second end tag was never detected in ct_netconfc:handle_data. Instead
it was included in the XML data given to the xmerl parser, which then
failed with reason "\"]]>\" is not allowed in content".

This problem was introduced by OTP-13007.