tftp: support transfer of files > 65535 blocks
The block count is an unsigned 2 byte integer. The common behaviour of
tftp clients and servers is to overflow the block count and wrap the
counter to 0 in order to support files larger than 32M. Modify the tftp
implementation to do the same rollover.

Interoperability was tested by transferring a 1.4G file using the HPA
tftp client.