Add issuer arg to ssl_crl_cache_api lookup callback
Change the ssl_crl_cache_api callback specification, passing the
certificate issuer name as an argument to the lookup callback
function.

The purpose of this change is to accomodate CRL cache modules that
index CRLs by issuer name, not by distribution point URL.

While in most cases such lookups could be performed using the select/2
callback function, that doesn't work when the CRL in question contains
an Issuing Distribution Point (IDP) extension, since RFC 5280
specifies different processing rules for CRLs specified in a
distribution point (DP) and other CRLs.  For the latter, a DP is
assumed that most likely will not match the IDP of the CRL.

In order to accommodate cache modules that index CRLs by issuer name,
let's pass them the issuer as well.