Add a -discontiguous directive as in EEP 38
This commit adds the directive

  -discontiguous([FunctionName/Arity, ...]).

"so that specified functions may be presented as more than one group of
clauses, possibly separated by other directives and function clause
groups." (quote from the EEP 38)

Thus, it is possible to write code like this (note how the clauses for
format_error are separated):

    -module(something).
    -compile([...]).
    -discontiguous([format_error/1]).

    some_function(...) -> ...
    format_error({bad_something,X}) -> ...

    some_other_function(...) -> ...
    format_error({invalid_other_thing,Y}) -> ...

There are more uses, see to the EEP for a more thorough
discussion.

This commit adds support for -discontiguous to the functions

   epp:parse_file/1,2,3 (and thus to compile:file)
   compile:forms/1,2
   compile:noenv_forms/2

The function compile:file eventually calls epp:parse_file, but it seems
reasonable that compiling equivalent forms to behave the same way. For
this to work, the new function

  epp:join_discontiguous_clauses(Forms) -> Forms1

has been exported.

As detailed in the EEP 38, upon parsing and joining discontiguous
functions, any processed -discontiguous directives are removed from the
resulting forms.