Produce warnings for binary patterns that will never match
Binary matching can be confusing. For example:

  1> <<-1>> = <<-1>>.
  ** exception error: no match of right hand side value <<"ÿ">>
  2>

When constructing binaries, the value will be masked to fit in
the binary segment. But no such masking happens when matching
binaries.

One solution that we considered was to do the same masking when
matching. We have rejected that solution for several reasons:

* Masking in construction is highly controversial and by some
  people considered a bad design decision.

* While masking of unsigned numbers can be understood, masking of
  signed numbers it not easy to understand.

* Then there is the question of backward compatibility. Adding
  masking to matching would mean that clauses that did not match
  earlier would start to match. That means that code that has
  never been tested will be executed. Code that has not been
  tested will usually not work.

Therefore, we have decided to warn for binary patterns that cannot
possibly match.

While we are it, we will also warn for the following example where
size for a binary segment is invalid:

  bad_size(Bin) ->
    BadSize = bad_size,
    <<42:BadSize>> = Bin.

That example would crash the HiPE compiler because the BEAM compiler
would generate a bs_get_integer2 instruction with an invalid size
field. We can avoid that crash if sys_core_fold not only warns for bad
binary pattern, but also removes the clauses that will not match.

Reported-by: http://bugs.erlang.org/browse/ERL-44
Reported-by: Kostis Sagonas