Remove the remove_event action and all alike
Removing events from the internal queues is not necessary with
the choosen semantics of the event queue vs. hibernate.

In an early implementation it was possible by combining
hibernate with e.g. postpone to get an event in the queue
that you would not see before processing the postponed event,
and therefore should you decide to cancel a timer it was
essential to be able to remove that unseen event from the queue.

With the choosen semantics you will have to postpone or generate
an event for it to be in the event queue, and if you e.g. postpone
a timeout event and then cancel the timer it is your mistake.
You have seen the event and should know better than to try to
cancel the timer.

So, the actions: remove_event, cancel_timer, demonitor and unlink
are now removed.

There have also been some cleanup of the timer handling code.