wx: Wait before throwing away early callbacks to wx_objects
If synced callbacks arrive early before the return value from
init had been synced to proc_lib. The callbacks where ignored.

Added a short timeout to avoid the scheduling raise but we don't want
block the wxe_server in case of deadlock in init, so sleep for a short
while and check again, and drop the callback when init takes to long.