asn1_SUITE_value: Don't test the LDAP module
asn1ct:test(Mod) will generate sample values for data types
in an ASN.1 modules. The following definition from LDAP
can cause problems for asn1ct:test/1:

  Filter ::= CHOICE {
          and             [0] SET SIZE (1..MAX) OF filter Filter,
          or              [1] SET SIZE (1..MAX) OF filter Filter,
          not             [2] Filter,
          equalityMatch   [3] AttributeValueAssertion,
          substrings      [4] SubstringFilter,
          greaterOrEqual  [5] AttributeValueAssertion,
          lessOrEqual     [6] AttributeValueAssertion,
          present         [7] AttributeDescription,
          approxMatch     [8] AttributeValueAssertion,
          extensibleMatch [9] MatchingRuleAssertion,
          ... }

If we were unlucky (which we were on some computers), the construction
of sample values would not terminate.

Eliminate the problem for now by not testing the LDAP module. This
workaround should be removed when we have added some limitation to the
recursion in asn1ct_value when generating values.