Add Ctrl+D support to ssh and user_drv
It works similar to the shell in Unix systems and
in many languages where Ctrl+D will shut the system
down if the current line is empty (i.e. no text before
nor after the current cursor position).

If there is text after, it will be deleted as previously.
If there is text before, a beep sound will be emitted.