io_SUITE: Don't fail on fast computers with rough timers
On a computer which is fast, but with timers with low resolution,
the measured time for an empty queue could be zero, which could
cause the test case to fail.

Add a calibration function to scale up the amount the amount of
work so that the measured time will not be lower than 50 ms.