Don't throw exceptions from rpc:call() and rpc:block_call()
The documentation for rpc:call() says:

  Evaluates apply(Module, Function, Args) on the node Node and
  returns the corresponding value Res, or {badrpc, Reason} if
  the call fails.

What is not said that rpc:call() can generate an exception if the
'rex' process on the other node is killed:

  (kalle@host)1> Rex = rpc:call(arne@host, erlang, whereis, [rex]).
  <6937.14.0>
  (kalle@host)2> rpc:call(arne@host, erlang, exit, [Rex,kill]).
  ** exception exit: {killed,
                       {gen_server,call,
                           [{rex,arne@host},
                            {call,erlang,exit,[<6937.14.0>,kill],<0.33.0>},
                            infinity]}}
       in function  rpc:rpc_check/1 (rpc.erl, line 361)

On the other hand, if the other node shuts down for some other reason,
we'll get a {badrpc,nodedown} result:

  (kalle@host)5> rpc:call(arne@host, erlang, halt, []).
  {badrpc,nodedown}

There does not seem to be any reason to handle the two cases
differently. If the 'rex' process is terminated on the other node,
it will shut down shortly thereafter.

Therefore, change rpc:call() and rpc:block_call() to always return
{badrpc,Reason} is the call fails:

  (kalle@host)1> Rex = rpc:call(arne@host, erlang, whereis, [rex]).
  <6937.14.0>
  (kalle@host)2> rpc:call(arne@host, erlang, exit, [Rex,kill]).
  {badrpc,{'EXIT',{killed,{gen_server,call,
                            [{rex,arne@host},
                             {call,erlang,exit,[<7126.14.0>,kill],<0.33.0>},
                                   infinity]}}}}