Add hook for node name check
Allow distribution carrier protocols to specify a function that checks
whether the name of the remote node is allowed.

This is useful when the carrier protocol authenticates the remote
node.  In that case, it needs to check that the node name matches the
presented credentials.

While it would be possible for the carrier protocol module to perform
this check by itself for outbound connections, it would still need
cooperation from dist_util to check node names for inbound
connections, since the node name is only revealed during the
handshake.  Thus, let's simplify by using the same mechanism in both
directions.