Make throttling callbacks on message reception
The callback is now applied to the atom 'false' when asking if another
message should be received on the socket, and to a received binary
message after reception. Throttling on received messages makes it
possible to distinguish between requests and answers.

There is no callback on outgoing messages since these don't have to go
through the transport process, even if they currently do.