configure.in: Fix compilation error in ethread.c
3121171f56c2 made warnings for missing prototypes into compilation
errors. That broke building on MacOS X, because prototypes for the
pthread_getname_np() and pthread_setname_np() functions included by
ethread.c are missing.

The prototypes for those functions *are* located in pthread.h,
but are not included if _XOPEN_SOURCE is defined. Therefore,
we must undefine _XOPEN_SOURCE.

6e606582a6 is the commit that started to define _XOPEN_SOURCE on
MacOS X in order to mend floating-point exceptions. Since
fp exceptions again are broken and disabled by default, it is
safe to remove the definition of _XOPEN_SOURCE.