Take out (parts of) broken fp exception support for MacOS X
Floating-point exception support on MacOS X has never been especially
reliable, and has therefore been disabled by default for a long time.

The fpe support is now broken.

Therefore, take out the unnecessary test for modern mcontext in
configure (whatever that means) and the associated code in sys_float.c.
Add #error directives to sys_float.c to make it clear that
fpe is not supported.

It seems to risky to mess with the mess of #ifdef's, so we will
not attempt to remove all fpe support code for MacOS X.