configure: Remove obsolete --enable-darwin-* options
Remove the options:

  --enable-darwin-universal
  --enable-darwin-64bit

The --enable-darwin-universal option turns on universal binaries
(Intel/PPC). It does not work on modern releases of OS X, and
OTP 19 will most not likely build on an OS X release that still
supports universal builds.

The --enable-darwin-64bit option is not needed, because 64-bit builds
are default on modern OS X releases. There is also the generic
--enable-m64-build option.