gen_sctp_SUITE: Skip most SCTP test cases on Solaris before 5.12
On old Solaris versions, many gen_sctp test cases fail. It is not
clear whether the implementation or the test cases or both must
be fixed.

The number of failed test cases is a problem, as other failed
test cases in the Kernel application might not be noticed.

For now, skip most test gen_sctp test cases on old Solaris platforms,
until the day that someone can mend the tests case and/or
implementation (or test machines die). Keep a few smoke test
cases.