reltool_server_SUITE: Correct diffing of start scripts
In c3c6182, systools was updated to add more modules to be pre-loaded
along with error_handler. That is a pure optimization.

Because of the extra modules, test cases that compared the start
scripts generated by SASL with the ones generated by Reltool
would fail.

We must implement a more elaborate way to diff the scripts.