gen_fsm_SUITE: Make abnormal1/1 stop failing
abnormal1/1 started failing sporadically when 22785cd1e21
changed using the messages_get/0 and using its own implementation:

get_messages() ->
    receive
	Msg -> [Msg|get_messages()]
    after 1 -> []
    end.

In test_server:messages_get/0, there was "after 0" instead of "after
1" as in this version.

We could change back to using "after 0", but it seems that the
test case just happened to work before the change.

Instead, read out the message that is guranteed to be in the message
queue, and *then* check that the message queue is empty.