Avoid potential timer bottleneck on supervisor restart
Prior to this patch, when the supervisor was unable to
restart a child, it used the timer process to execute
a function later on, allowing other messages in the
supervisor queue to be consumed.

This patch keeps the same idea but bypasses the timer
process by directly sending a message to the supervisor
process itself.