Simplify the raise instruction to reduce code size
The raise/2 instruction is almost always used like this:

  raise x(2) x(1)

Therefore, we can translate it to an internal i_raise/0
instruction that uses x(2) x(1) as its implicit operands.

We will also remove the backward compatibility with R10-0.  It is
unlikely that anyone still is using BEAM files compiled with the R10-0
compiler, especially since most of those modules cannot be loaded. The
loader will refuse to load any module that uses the old non-GCIng
arithmetic instructions or the non-GCing versions of length/1 or
size/1.

Doing these changes will reduce both the size of the loaded BEAM
code and size of the code in process_main().