erts: Optimize '++' operator
This also optimizes the BIF lists:append/2

Use one pass to check for properness and copying LHS list.
If LHS turns out not being a proper list, bail and reset htop.
If we run out of heap, allocate a heap-fragment and calculate
the remaining length as normal, thus checking for properness,
and then continue copying.

Measurements shows this being ~50% faster.