v3_kernel: Construct literal lists properly
Use cerl:make_list/1 instead of a home-made make_list/1 to ensure that
literal lists are constructed as literals. In a future release, we
would like to forbid in the loader construction of literal lists using
instructions like:

   put_list {atom,a} [] Dst

The proper way is:

   move {literal,[a]} {x,0}

Also update the comment about "put_list Const [] Dst" in ops.tab.