Correct unpacking of 3 operands on 32-bit archictectures
0a4750f91c83 optimized unpacking by removing a mask operation
when unpacking three packed operands. Unfortunately, that optimization
is only safe on 64-bit architectures.

Here is what happens on 32-bit architectures.

The operands to be packed are 10-bit register numbers that have been
turned into byte offsets:

  aaaaaaaaaa00
  bbbbbbbbbb00
  cccccccccc00

They can be packed into a single word like this:

    30         20         10        0
     |          |          |        |
  aa aaaaaaaabb bbbbbbbbcc cccccccc00

If we call the packed word P, the original operands can be
extracted like this:

  C =  P         band 2#111111111100
  B = (P bsr 10) band 2#111111111100
  A = (P bsr 20) band 2#111111111100

The bug was that A was extracted without the masking:

  A = P bsr 20

That would give A the value:

  aaaaaaaaaaaabb

That would only be safe if the two most significant bits in B
were zeroes.