Remove timestamps from BEAM files
As long as anyone can remember, the compilation time has been included
in BEAM files (and can be retrieved using Mod:module_info(compile)).

The timestamp has caused problems for anyone using tools such as 'cmp'
to compare BEAM files or for package managers:

  http://erlang.org/pipermail/erlang-questions/2016-April/088717.html

Rarely has the timestamp been of any use. Yes, sometimes the timestamp
could help to figuring out which version of a module was used, but
nowadays a better way is to use Mod:module_info(md5).

To get rid of this problem, remove the timestamp from BEAM files in
OTP 19. Don't add an option to include timestamps.

Utilities that depend on the timestamp will need to be modified.
For example:

  http://erlang.org/pipermail/erlang-questions/2016-April/088730.html

Instead of using the compilation time, the MD5 for the BEAM code can
be used. Example:

1> c:module_info(md5).
<<79,26,188,243,168,60,58,45,34,69,19,222,138,190,214,118>>
2> beam_lib:md5(code:which(c)).
{ok,{c,<<79,26,188,243,168,60,58,45,34,69,19,222,138,190,214,118>>}}
3>