Move bit syntax test cases from compilation_SUITE to bs_match_SUITE
We used to put code that would crash the compiler into
compilation_SUITE_data. That way we would have a failing test case to
remind us to fix a bug.

Nowadays, we generally fix the bug and write the test case at the same
time. Therefore it makes more sense to put the test code directly into
a test suite.

Move out bin_syntax_1 through bin_syntax_5 test cases. Scrap
bin_syntax_6 because it does not longer seems to be relevant.

While we are it, rename the fun_shadow/1 test to size_shadow/1. Also
make sure that the code produces the correct result.